% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_densities.R
\name{calculate_densities}
\alias{calculate_densities}
\title{Calculate Densities}
\usage{
calculate_densities(x, y, size, shift = NULL)
}
\arguments{
\item{x}{RasterStack with realizations}

\item{y}{RasterStack with shares}

\item{size}{size of a motifel}

\item{shift}{shift of a motifel}
}
\value{
a list of data.frames
}
\description{
Calculate Densities
}
\examples{
\dontrun{
library(raster)
real_raster = create_realization(perc_raster)
plot(real_raster)
b = calculate_density(real_raster, perc_raster, size = 10)

real_rasters = create_realizations(perc_raster, n = 5)
plot(real_rasters)
d = calculate_densities(real_rasters, perc_raster, size = 10)
}
}
