% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Title}
\usage{
calculate_metrics(x, w, neighbourhood, fun, size = NULL,
  na_action = "replace", base = "log2", ordered = TRUE)
}
\arguments{
\item{ordered}{}
}
\description{
Title
}
\examples{
library(sd)
library(raster)
x = create_realization(perc_raster)
w = create_weights(x, perc_raster, size = 100)
df = calculate_metrics(x, w, neighbourhood = 4, fun = "mean")
df = calculate_metrics(x, w, neighbourhood = 4, fun = "mean", size = 100)
}
