% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weights.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Title}
\usage{
calculate_weights(x, y, size)
}
\arguments{
\item{size}{}
}
\description{
Title
}
\examples{
a = create_realization(perc_raster)
plot(a)
}
