% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_densities.R
\name{create_densities}
\alias{create_densities}
\title{Create Densities}
\usage{
create_densities(x, y, window_size)
}
\arguments{
\item{x}{RasterStack with realizations}

\item{y}{RasterStack with shares}

\item{window_size}{size of a local window}
}
\value{
a RasterStack
}
\description{
Create Densities
}
\examples{
library(raster)
real_rasters = create_realizations(race_raster, n = 5)
plot(real_rasters)
dens_raster = create_densities(real_rasters, race_raster, window_size = 10)
dens_raster
plot(dens_raster)

}
