% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_realizations.R
\name{create_realizations}
\alias{create_realizations}
\title{Title}
\usage{
create_realizations(x, n)
}
\arguments{
\item{x}{RasterStack/RasterBrick}

\item{n}{Number of realizations}
}
\description{
Title
}
\examples{
\dontrun{
system.time({a = create_realization(perc_raster)})
plot(a)
system.time({b = create_realizations(perc_raster, 10)})
plot(b)
}
}
