% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_weights.R
\name{create_weights}
\alias{create_weights}
\title{Title}
\usage{
create_weights(x, y, size)
}
\arguments{
\item{size}{}
}
\description{
Title
}
\examples{
\dontrun{
library(raster)
real_raster = create_realization(perc_raster)
plot(real_raster)
output = create_weights(real_raster, perc_raster, size = 10)
output
}
}
