% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raceland.R
\name{get_raceland}
\alias{get_raceland}
\title{Title}
\usage{
get_raceland(
  x,
  n,
  window_size,
  neighbourhood,
  fun,
  size = NULL,
  shift = NULL,
  na_action = "replace",
  base = "log2",
  ordered = TRUE,
  threshold = 0.5
)
}
\arguments{
\item{threshold}{}
}
\value{

}
\description{
Title
}
\examples{
library(raster)
rl = get_raceland(race_raster, 10)
}
