% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_viz.R
\name{my_viz}
\alias{my_viz}
\title{Title}
\usage{
my_viz(race_raster, real_raster, hex)
}
\arguments{
\item{race_raster}{A}

\item{real_raster}{B}

\item{hex}{C}
}
\value{
D
}
\description{
Title
}
\examples{
library(raster)
hex_colors = c("#F16667", "#6EBE44", "#7E69AF", "#C77213","#F8DF1D")
real = create_realizations(race_raster, 1)
plot(race_raster)
plot(real)

my_viz(race_raster, real, hex_colors)
}
