% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_vector}
\alias{pop_vector}
\title{An sf object}
\format{An sf object}
\usage{
pop_vector
}
\description{
It is an sf POLYGON object with census block-level data. It consists of 7 variables: GISJOIN - block ID, ASIAN, BLACK, HISPANIC, OTHER, WHITE - number of people of given race/ethnicity in each block
}
\keyword{datasets}
