% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{race_raster}
\alias{race_raster}
\title{A RasterStack}
\format{A RasterLayer}
\usage{
race_raster
}
\description{
A RasterStack covering an area of 60x60 cells. RasterStack contains 5 layers - a high resolution (30m) race-specific grids with values of subpopulation densities for Asian, Black, Hispanic, other and Whites
}
\keyword{datasets}
