% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{race_raster.tif}
\alias{race_raster.tif}
\title{A raster file}
\format{
A raster file
}
\description{
A raster file covering an area of 60x60 cells. The raster file contains 5 layers - a high resolution (30m) race-specific grids with values of subpopulation densities for Asian, Black, Hispanic, other and Whites.
\code{system.file("extdata/race_raster.tif", package = "raceland")}
}
