% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceland-package.R
\docType{package}
\name{raceland-package}
\alias{raceland}
\alias{raceland-package}
\title{raceland: Pattern-based Zoneless Method for Analysis and Visualization of Racial Topography}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (0000-0002-1057-3721)

Authors:
\itemize{
  \item Anna Dmowska
  \item Tomasz Stepinski
}

}
\keyword{internal}
