% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_densities.R
\name{calculate_densities}
\alias{calculate_densities}
\title{Calculate Densities}
\usage{
calculate_densities(x, y, window_size)
}
\arguments{
\item{x}{SpatRaster with realizations}

\item{y}{SpatRaster with shares}

\item{window_size}{size of a local window}
}
\value{
a list of data.frames
}
\description{
Calculate Densities
}
\examples{
library(terra)
race_raster = rast(system.file("extdata/race_raster.tif", package = "raceland"))
real_rasters = create_realizations(race_raster, n = 5)
plot(real_rasters)
d = raceland:::calculate_densities(real_rasters, race_raster, window_size = 10)

}
\keyword{internal}
