% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate Metrics}
\usage{
calculate_metrics(x, w, neighbourhood, fun, size = NULL, shift = NULL,
  na_action = "replace", base = "log2", ordered = TRUE)
}
\arguments{
\item{x}{RasterStack with realizations}

\item{w}{RasterStack with weights}

\item{neighbourhood}{4/8}

\item{fun}{"mean", "geometric_mean", or "focal"}

\item{size}{size of a motifel}

\item{shift}{shift of a motifel}

\item{na_action}{"replace", "omit", "keep"}

\item{base}{the unit in which entropy is measured. The default is "log2", which compute entropy in "bits". "log" and "log10" can be also used}

\item{ordered}{the type of pairs considered. Either ordered (TRUE) or unordered (FALSE). The default is TRUE}
}
\value{
a data.frame
}
\description{
Calculate Metrics
}
\examples{
library(racialscape)
library(raster)
x = create_realizations(perc_raster, n = 5)
w = create_weights(x, perc_raster, size = 100)

#1
df = calculate_metrics(x, w, neighbourhood = 4, fun = "mean")

#2
df2 = calculate_metrics(x, w, neighbourhood = 4, fun = "mean", size = 100)
my_grid = create_grid(x, size = 100)
\dontrun{
 df3 = dplyr::filter(df2, realization == 2)
 result = dplyr::left_join(my_grid, df2, by = c("row", "col"))
 plot(result)
}
}
