% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_weights.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Title}
\usage{
calculate_weights(x, y, size)
}
\arguments{
\item{size}{}
}
\description{
Title
}
\examples{
\dontrun{
library(raster)
real_raster = create_realization(perc_raster)
plot(real_raster)
b = calculate_weights(real_raster, perc_raster, size = 10)
}
}
