% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Title}
\usage{
create_grid(x, size)
}
\arguments{
\item{size}{}
}
\description{
Title
}
\examples{
x = create_realization(perc_raster)
y = create_grid(x, size = 10)
y
}
