% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_realizations.R
\name{create_realizations}
\alias{create_realizations}
\title{Create Realizations}
\usage{
create_realizations(x, n)
}
\arguments{
\item{x}{RasterStack with shares}

\item{n}{number of realizations}
}
\value{
a RasterStack
}
\description{
Create Realizations
}
\examples{
library(raster)
real = create_realizations(race_raster, 10)
plot(real)

}
