% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raceland-package.R
\docType{package}
\name{raceland-package}
\alias{raceland}
\alias{raceland-package}
\title{raceland: Pattern-Based Zoneless Method for Analysis and Visualization of Racial Topography}
\description{
Implements a computational framework for a pattern-based, 
    zoneless analysis, and visualization of (ethno)racial topography. It is a reimagined
    approach for analyzing residential segregation and racial diversity based on 
    the concept of 'landscape’ used in the domain of landscape ecology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nowosad.github.io/raceland/}
  \item Report bugs at \url{https://github.com/Nowosad/raceland/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

Authors:
\itemize{
  \item Anna Dmowska
  \item Tomasz Stepinski
}

}
\keyword{internal}
