% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-package.R
\docType{package}
\name{sd-package}
\alias{sd}
\alias{sd-package}
\title{sd: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (0000-0002-1057-3721)

Authors:
\itemize{
  \item Anna Dmowska
  \item Tomasz Stepinski
}

}
\keyword{internal}
